# src/integration-test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: care-management
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:63933
  description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/errands/{errandId}/attachments:
    get:
      tags:
      - Errand attachments
      summary: List attachments
      description: "Fetches all attachment metadata for the errand (application files,\
        \ generated PDFs, manual uploads and conversation attachments), oldest first.\
        \ Optionally filtered by origin and/or sender role."
      operationId: readAttachments
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      - name: origin
        in: query
        description: Only return attachments with this origin
        required: false
        schema:
          type: string
          enum:
          - APPLICATION
          - CONVERSATION
          - GENERATED
          - ERRAND
      - name: senderRole
        in: query
        description: Only return attachments from this sender
        required: false
        schema:
          type: string
          enum:
          - CLIENT
          - CASEWORKER
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Errand attachments
      summary: Create attachment
      description: Uploads a new attachment for the errand
      operationId: createAttachment
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}/messages/{messageId}/attachments/{attachmentId}/file:
    get:
      tags:
      - Conversation
      summary: Download a message attachment
      description: Streams the attachment file content
      operationId: streamAttachmentFile
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: messageId
        in: path
        required: true
        schema:
          type: string
      - name: attachmentId
        in: path
        description: Attachment id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}/attachments/{attachmentId}:
    get:
      tags:
      - Errand attachments
      summary: Read attachment metadata
      description: Fetches the attachment metadata
      operationId: readAttachment
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      - name: attachmentId
        in: path
        description: Attachment id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Errand attachments
      summary: Delete attachment
      description: Deletes the attachment
      operationId: deleteAttachment
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      - name: attachmentId
        in: path
        description: Attachment id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{namespace}/errands/{errandId}/attachments/{attachmentId}/file:
    get:
      tags:
      - Errand attachments
      summary: Download attachment file
      description: Streams the attachment file content
      operationId: streamAttachmentFile_1
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      - name: attachmentId
        in: path
        description: Attachment id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Attachment:
      type: object
      description: Attachment model
      properties:
        id:
          type: string
          description: Unique identifier
          examples:
          - cb20c51f-fcf3-42c0-b613-de563634a8ec
          readOnly: true
        fileName:
          type: string
          description: File name
          examples:
          - my-file.pdf
        mimeType:
          type: string
          description: Mime type
          examples:
          - application/pdf
        fileSize:
          type: integer
          format: int32
          description: File size in bytes
          examples:
          - 1024
          readOnly: true
        origin:
          type: string
          description: "Where the file came from: APPLICATION (citizen's application\
            \ files), CONVERSATION (sent in a message thread), GENERATED (a consolidated\
            \ PDF produced by the platform) or ERRAND (uploaded directly to the errand)"
          enum:
          - APPLICATION
          - CONVERSATION
          - GENERATED
          - ERRAND
          examples:
          - CONVERSATION
          readOnly: true
        senderRole:
          type: string
          description: "Who the file came from: CLIENT (applicant) or CASEWORKER (caseworker).\
            \ May be null for files predating the distinction or with no clear sender."
          enum:
          - CLIENT
          - CASEWORKER
          examples:
          - CLIENT
          readOnly: true
        messageId:
          type: string
          description: "For CONVERSATION attachments, the id of the message the file\
            \ is attached to — download it via .../messages/{messageId}/attachments/{id}/file.\
            \ Null for non-conversation attachments, which download via .../attachments/{id}/file."
          readOnly: true
        created:
          type: string
          format: date-time
          description: Created timestamp
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Modified timestamp
          readOnly: true
...
  securitySchemes: {}